requirements.txt
# VORTEX - Deep Research Agent Dependencies
# ==========================================

# LangGraph & LangChain Core
langgraph
langchain
langchain_openai

# Research Tools
tavily-python

# Data Validation
pydantic

# Environment Management
python-dotenv

# Testing & Development
pytest
jupyter
langchain-community

# RAG Integration
httpx
langchain-openai
